Separate values and initialValues from field state
Turns out that storing the values and state of fields in the same tree isn’t a very good idea. This makes it very hard to set the value of an entire subtree, instead of just updating the values of leafs.

This also solves the issue where setting an initial value for a field caused an extra initial render.